# docs/openapi.yaml
openapi: 3.1.0
info:
  title: KonaMall API
  description: |
    글로벌 직구 드롭쉬핑 플랫폼 KonaMall의 REST API 문서입니다.
    
    ## 인증
    JWT Bearer 토큰을 사용합니다. 로그인 후 받은 `access_token`을 
    `Authorization: Bearer {token}` 헤더에 포함하여 요청하세요.
    
    ## 에러 응답
    모든 에러는 다음 형식으로 반환됩니다:
    ```json
    {
      "detail": "에러 메시지"
    }
    ```
  version: 1.0.0
  contact:
    name: KonaMall Support
    email: support@konamall.com
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.konamall.com
    description: Production
...
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: 상품 목록 조회
      operationId: listProducts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          description: 검색어 (제목, 설명)
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: supplier_id
          in: query
          schema:
            type: integer
        - name: min_price
          in: query
          description: 최소 가격 (KRW)
          schema:
            type: integer
        - name: max_price
          in: query
          description: 최대 가격 (KRW)
          schema:
            type: integer
        - name: sort
          in: query
          description: 정렬 기준
          schema:
            type: string
            enum: [price_asc, price_desc, newest, popular]
            default: newest
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
  /api/products/{product_id}:
    get:
      tags: [Products]
      summary: 상품 상세 조회
      operationId: getProduct
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '404':
          description: 상품 없음

  # ========== Cart ==========
  /api/cart:
    get:
      tags: [Cart]
      summary: 장바구니 조회
      operationId: getCart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
    delete:
      tags: [Cart]
      summary: 장바구니 비우기
      operationId: clearCart
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 성공
  /api/cart/items:
    post:
      tags: [Cart]
      summary: 장바구니에 상품 추가
      operationId: addCartItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemAdd'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: 재고 부족
  /api/cart/items/{item_id}:
    parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [Cart]
      summary: 장바구니 아이템 수량 변경
      operationId: updateCartItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemUpdate'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
    delete:
      tags: [Cart]
      summary: 장바구니에서 상품 제거
      operationId: removeCartItem
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 성공

  # ========== Orders ==========
  /api/orders:
    get:
      tags: [Orders]
      summary: 내 주문 목록
      operationId: listMyOrders
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
    post:
      tags: [Orders]
      summary: 주문 생성 (장바구니 → 주문)
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: 주문 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: 장바구니가 비어있거나 재고 부족
  /api/orders/{order_id}:
    get:
      tags: [Orders]
      summary: 주문 상세 조회
      operationId: getOrder
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '404':
          description: 주문 없음
...
  /api/admin/suppliers:
    get:
      tags: [Suppliers]
      summary: 공급자 목록 (Admin)
      operationId: listSuppliers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Supplier'
    post:
      tags: [Suppliers]
      summary: 공급자 추가 (Admin)
      operationId: createSupplier
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierCreate'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
  /api/admin/suppliers/{supplier_id}/sync:
    post:
      tags: [Suppliers]
      summary: 상품 동기화 시작 (Admin)
      operationId: syncSupplierProducts
      security:
        - bearerAuth: []
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '202':
          description: 동기화 작업 시작됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  task_id:
                    type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        title_ko:
          type: string
        price_original:
          type: number
          description: USD
        price_final:
          type: integer
          description: KRW (마진 포함)
        currency:
          type: string
        stock:
          type: integer
        main_image:
          type: string
          format: uri
        supplier_name:
          type: string
        category:
          type: string
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            description:
              type: string
            description_ko:
              type: string
            images:
              type: array
              items:
                $ref: '#/components/schemas/ProductImage'
            variants:
              type: array
              items:
                $ref: '#/components/schemas/ProductVariant'
            shipping_days_min:
              type: integer
            shipping_days_max:
              type: integer
            origin_url:
              type: string
              format: uri
    ProductImage:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
          format: uri
        is_main:
          type: boolean
    ProductVariant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
        price_krw:
          type: integer
        stock:
          type: integer
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer

    # ========== Cart ==========
    Cart:
      type: object
      properties:
        id:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          type: integer
          description: 소계 (KRW)
        shipping_fee:
          type: integer
        total:
          type: integer
    CartItem:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        variant_id:
          type: integer
        product:
          $ref: '#/components/schemas/Product'
        variant:
          $ref: '#/components/schemas/ProductVariant'
        quantity:
          type: integer
        line_total:
          type: integer
    CartItemAdd:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: integer
        variant_id:
          type: integer
        quantity:
          type: integer
          minimum: 1
          default: 1
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1

    # ========== Order ==========
    OrderStatus:
      type: string
      enum: [pending, paid, processing, shipped, delivered, cancelled, refunded]
    PaymentStatus:
      type: string
      enum: [pending, completed, failed, refunded]
    Order:
      type: object
      properties:
        id:
          type: integer
        order_number:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        total_amount:
          type: integer
        items_count:
          type: integer
        created_at:
          type: string
          format: date-time
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'
            subtotal_krw:
              type: integer
            shipping_cost_krw:
              type: integer
            shipping_name:
              type: string
            shipping_phone:
              type: string
            shipping_address:
              type: string
            payment_method:
              type: string
            paid_at:
              type: string
              format: date-time
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product_title:
          type: string
        variant_name:
          type: string
        quantity:
          type: integer
        unit_price:
          type: integer
        image_url:
          type: string
    OrderCreate:
      type: object
      required: [address_id]
      properties:
        address_id:
          type: integer
          description: 배송지 ID
        payment_method:
          type: string
          enum: [card, bank_transfer, kakao_pay, naver_pay]
        note:
          type: string
          description: 배송 메모
    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ========== Shipment ==========
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        supplier_type:
          $ref: '#/components/schemas/SupplierType'
        is_active:
          type: boolean
        products_count:
          type: integer
    SupplierType:
      type: string
      enum: [temu, aliexpress, amazon, local]
    SupplierCreate:
      type: object
      required: [name, supplier_type]
      properties:
        name:
          type: string
        supplier_type:
          $ref: '#/components/schemas/SupplierType'
        base_url:
          type: string
        api_key:
          type: string
        api_secret:
          type: string